NOW ADD TO RING WORKS AND WE GET PROPPER PRINTOUTS!!!!
Eka helped me understand the difference between sending in a whole node
to the addtoring and just sending a pointer. Hade the problem with
sending and making a copy of a node instead of just sending the
pointer. Here we can se the difference of what a * can do to a system